requirements.txt
# Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.0
requests>=2.31.0

# AI Analysis
openai>=1.0.0

# Configuration
python-dotenv>=1.0.0

# Web Application
django>=4.2
django-apscheduler>=0.6.2
